Remove deprecated FRAPI fallback consumers
- Remove appropriate default method implementations
- Move some common code between AbstractBlockRenderContext and ItemRenderContext into AbstractRenderContext
