+++ fabric-renderer-api-v1/src/client/java/net/fabricmc/fabric/api/renderer/v1/material/MaterialView.java
@@ -57,10 +57,6 @@ public interface MaterialView {
 
 	/**
 	 * @see MaterialFinder#shadeMode(ShadeMode)
-	 *
-	 * @apiNote The default implementation will be removed in the next breaking release.
 	 */
-	default ShadeMode shadeMode() {
-		return ShadeMode.ENHANCED;
-	}
+	ShadeMode shadeMode();
 }
