+++ fabric-renderer-api-v1/src/client/java/net/fabricmc/fabric/impl/renderer/VanillaModelEncoder.java
@@ -38,7 +38,6 @@ import net.fabricmc.fabric.api.util.TriState;
 
 /**
  * Routines for adaptation of vanilla {@link BakedModel}s to FRAPI pipelines.
- * Even though Indigo calls them directly, they are not for use by third party renderers, and might change at any time.
  */
 public class VanillaModelEncoder {
 	private static final Renderer RENDERER = RendererAccess.INSTANCE.getRenderer();
