+++ fabric-renderer-indigo/src/client/java/net/fabricmc/fabric/impl/client/indigo/renderer/render/AbstractBlockRenderContext.java
@@ -27,40 +27,22 @@ import net.minecraft.client.render.LightmapTextureManager;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.VertexConsumer;
 import net.minecraft.client.render.WorldRenderer;
-import net.minecraft.client.render.model.BakedModel;
 import net.minecraft.client.render.model.json.ModelTransformationMode;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
 
 import net.fabricmc.fabric.api.renderer.v1.material.RenderMaterial;
 import net.fabricmc.fabric.api.renderer.v1.material.ShadeMode;
-import net.fabricmc.fabric.api.renderer.v1.mesh.QuadEmitter;
 import net.fabricmc.fabric.api.util.TriState;
 import net.fabricmc.fabric.impl.client.indigo.Indigo;
 import net.fabricmc.fabric.impl.client.indigo.renderer.aocalc.AoCalculator;
 import net.fabricmc.fabric.impl.client.indigo.renderer.aocalc.AoConfig;
 import net.fabricmc.fabric.impl.client.indigo.renderer.helper.ColorHelper;
-import net.fabricmc.fabric.impl.client.indigo.renderer.mesh.EncodingFormat;
 import net.fabricmc.fabric.impl.client.indigo.renderer.mesh.MutableQuadViewImpl;
-import net.fabricmc.fabric.impl.renderer.VanillaModelEncoder;
 
 public abstract class AbstractBlockRenderContext extends AbstractRenderContext {
 	protected final BlockRenderInfo blockInfo = new BlockRenderInfo();
 	protected final AoCalculator aoCalc;
-
-	private final MutableQuadViewImpl editorQuad = new MutableQuadViewImpl() {
-		{
-			data = new int[EncodingFormat.TOTAL_STRIDE];
-			clear();
-		}
-
-		@Override
-		public void emitDirectly() {
-			renderQuad(this);
-		}
-	};
-
-	private final BakedModelConsumerImpl vanillaModelConsumer = new BakedModelConsumerImpl();
 
 	private final BlockPos.Mutable lightPos = new BlockPos.Mutable();
 
@@ -73,12 +55,6 @@ public abstract class AbstractBlockRenderContext extends AbstractRenderContext {
 	protected abstract VertexConsumer getVertexConsumer(RenderLayer layer);
 
 	@Override
-	public QuadEmitter getEmitter() {
-		editorQuad.clear();
-		return editorQuad;
-	}
-
-	@Override
 	public boolean isFaceCulled(@Nullable Direction face) {
 		return !blockInfo.shouldDrawFace(face);
 	}
@@ -89,11 +65,7 @@ public abstract class AbstractBlockRenderContext extends AbstractRenderContext {
 	}
 
 	@Override
-	public BakedModelConsumer bakedModelConsumer() {
-		return vanillaModelConsumer;
-	}
-
-	private void renderQuad(MutableQuadViewImpl quad) {
+	protected void renderQuad(MutableQuadViewImpl quad) {
 		if (!transform(quad)) {
 			return;
 		}
@@ -272,26 +244,5 @@ public abstract class AbstractBlockRenderContext extends AbstractRenderContext {
 
 		// Unfortunately cannot use brightness cache here unless we implement one specifically for flat lighting. See #329
 		return WorldRenderer.getLightmapCoordinates(blockInfo.blockView, blockState, lightPos);
-	}
-
-	/**
-	 * Consumer for vanilla baked models. Generally intended to give visual results matching a vanilla render,
-	 * however there could be subtle (and desirable) lighting variations so is good to be able to render
-	 * everything consistently.
-	 *
-	 * <p>Also, the API allows multi-part models that hold multiple vanilla models to render them without
-	 * combining quad lists, but the vanilla logic only handles one model per block. To route all of
-	 * them through vanilla logic would require additional hooks.
-	 */
-	private class BakedModelConsumerImpl implements BakedModelConsumer {
-		@Override
-		public void accept(BakedModel model) {
-			accept(model, blockInfo.blockState);
-		}
-
-		@Override
-		public void accept(BakedModel model, @Nullable BlockState state) {
-			VanillaModelEncoder.emitBlockQuads(model, state, blockInfo.randomSupplier, AbstractBlockRenderContext.this);
-		}
 	}
 }
